+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ In Development
 ~~~~~~~~~~~~~~
 
 * Add HTTP host header whitelist (`!2 <https://gitlab.com/tmkn/django-basic-auth-ip-whitelist/merge_requests/2>`_)
+* Add realm setting (`!3 <https://gitlab.com/tmkn/django-basic-auth-ip-whitelist/merge_requests/3>`_)
 
 0.1 - Initial release
 ~~~~~~~~~~~~~~~~~~~~~